# Speed and performance improvement and Last commit fix