Speed and performance improvement and Last commit fix
1. Speed improvement: optimization of Ewald Fourier wave vector initialization
  * Example speed improvement: (Performed on RTX 4090)
    * CO2-MFI: 9.7 seconds to 9.0 seconds
    * NU-2000: 17.3 seconds to 11.6 seconds
2. Last commit changed function names, but pushed the wrong files
  * fixed